monty-backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.4.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.10.4
attrs==25.4.0
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
distro==1.9.0
ecdsa==0.19.1
fastapi==0.109.0
frozenlist==1.8.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.11
jiter==0.13.0
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.1
openai==2.21.0
passlib==1.7.4
propcache==0.4.1
psycopg2-binary==2.9.9
pyasn1==0.6.2
pycparser==3.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dotenv==1.0.0
python-jose==3.3.0
python-telegram-bot==21.0
pytz==2025.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.35.1
tqdm==4.67.3
typing_extensions==4.15.0
tzlocal==5.3.1
uvicorn==0.27.0
yarl==1.22.0
